+++ src/lib/zodSchemas.ts
@@ -5,6 +5,8 @@ import {
   equipments,
   exerciseEventFieldsToShow,
   exerciseTypes,
+  grades,
+  gradeSystems,
   muscleGroups,
   muscles,
   postures,
@@ -153,3 +155,46 @@ export type ProfileSchema = typeof profileSchema;
 
 export const profilePartialSchema = profileSchema.partial();
 export type ProfilePartialSchema = typeof profilePartialSchema;
+
+const projectBaseSchema = z.object({
+  name: z.string().min(1),
+  notes: z.string().nullish(),
+  grade: z.string(),
+  gradeSystem: z.string(),
+  url: z.string().nullish(),
+  imageS3ObjectKey: z.string().nullish(),
+});
+export const projectPartialBaseSchema = projectBaseSchema.partial();
+
+const projectSchemaChecks = (
+  val: z.infer<typeof projectBaseSchema> | z.infer<typeof projectPartialBaseSchema>,
+  ctx: z.RefinementCtx
+) => {
+  if (val.gradeSystem && !gradeSystems.includes(val.gradeSystem)) {
+    ctx.addIssue({
+      code: z.ZodIssueCode.custom,
+      message: `Invalid grade system`,
+      path: ['gradeSystem'],
+    });
+  }
+  if (val.gradeSystem && val.grade && !grades[val.gradeSystem].includes(val.grade)) {
+    ctx.addIssue({
+      code: z.ZodIssueCode.custom,
+      message: `Grade is required and must be a valid grade for the selected system`,
+      path: ['grade'],
+    });
+  }
+};
+
+export const projectSchema = projectBaseSchema.superRefine(projectSchemaChecks);
+export type ProjectSchema = typeof projectSchema;
+
+export const projectPartialSchema = projectPartialBaseSchema.superRefine(projectSchemaChecks);
+export type ProjectPartialSchema = typeof projectPartialSchema;
+
+export const projectSessionSchema = z.object({
+  notes: z.string().nullish(),
+  date: z.date(),
+  sent: z.boolean().default(false).optional(),
+});
+export type ProjectSessionSchema = typeof projectSessionSchema;
